move checkers out of .d.ts files
